requirements.txt
Flask
requests
bs4
tldextract
whois
dnspython
gunicorn
pyOpenSSL